Created cs files for Exception assignment
I created a list of integers and had the user input a number that is used in a foreach loop to divide each list element by that number. I put the loop in a try block and made two catch blocks to appropriate error messages when the user inputs 0 or a data type that can't be converted to an integer. I also included a finally block to show the results of the loop and the catch blocks.