I created the cs files for the Calling Methods assignment
I created a class with three methods that took a number input by the user and performed math operations on it. I then called each method twice: once to perform each method's operation on the number, returning all three results, and second to perform all three method's operations on the number in sequence before returning a single result to the user in the console.