Created cs files for Parameters assignment
I assigned a generic data type to the Employee class and to its List parameter. I then instantiated two employee objects, one with a string data type and the other with an int data type. I then assigned values to each list and made a loop to print out each value to the console.