Created cs file for Parsing Enums assignment
I created an enum for the days of the week and asked the user to input the current day. I then ran an if statement to check if the value could be converted to an int. If it could then , it sends a message to the user to enter a day of the week (I did this because entering numbers 0-6 also related to the enum values). If it couldn't be converted to an integer, then it parses the answer to the DayOfTheWeek enum and prints the results to the console. The entire set of code is set in a try catch in case the user misspells the day or enters something other than the specified weekday names or an integer. Should the user enter another value, the catch block tells them to enter a day of the week.