Created cs file for the DateTime assignment
I created a file that retrieves and prints the current DateTime to the console and then asks the user to enter a number. That number is then calculated into a TimeSpan as hours and added to the current DateTime and printed to the console to show the exact DateTime that many hours later.